Move `peft` out of core dependencies
`peft` requires `torch` as a mandatory dependency, so adding `peft` as a
core dependency means that `torch` will be installed i.e. even when `pip
install -e ".[quality]"` which slows things a bit, since `torch`
installation takes time; and in some cases is not needed, so moving
`peft` into the `torch` extra makes the most sense for the moment (not
final).